split braille test and method pass
