+++ lib/translator_to_english.rb
@@ -3,4 +3,9 @@ require './lib/file_reader'
 
 class TranslatorToEnglish
   include BrailleDictionary
+
+  def split_braille(message)
+    length = message.length / 3
+    message.scan(/.{1,#{length}}/)
+  end
 end
